chore(scripts): move federation e2e scripts to deferred/ + correct relative paths
Five e2e scripts (e2e-federation, e2e-pom, e2e-revocation, e2e-autosync,
e2e-federation-network) imported `../dist/services/federation.js`, which
is build-deferred per `mcp-server/tsconfig.json` (`"exclude": ["src/deferred/**"]`).
Running any of them today crashes with `ERR_MODULE_NOT_FOUND` for federation.js.

Commit b293040 ("chore: refocus mycelium on the brain core, decouple openClaw")
moved e2e-push-pom.mjs to mcp-server/scripts/deferred/ to mirror the source
placement, but left the rest behind. Complete the move so the directory
structure tells the truth: federation-dependent e2e scripts live next to the
deferred services they exercise, and stay inert until the deferred build is
unparked (CLAUDE.md roadmap step 5).

Side effect: e2e-push-pom.mjs's `../dist/...` paths were broken after its move
(resolves to mcp-server/scripts/dist instead of mcp-server/dist). Bump every
script in this directory to `../../dist/...` so the imports resolve correctly
once the federation build is re-emitted. Verified locally: e2e-pom now errors
on `dist/services/federation.js` (the deferred-build cause) instead of the
path mismatch — IdentityService and GuardService imports resolve cleanly.

Build green (`npm run build`). No active script or package.json target
references these files; they are manually-invoked diagnostic tools.
