+++ mcp-server/scripts/deferred/e2e-push-pom.mjs
@@ -10,9 +10,9 @@
 //      → cert mismatch detected (simulate via direct challengePom mit falscher expected pubkey)
 
 import { execSync } from "node:child_process";
-import { IdentityService } from "../dist/services/identity.js";
-import { FederationService } from "../dist/services/federation.js";
-import { GuardService } from "../dist/services/guard.js";
+import { IdentityService } from "../../dist/services/identity.js";
+import { FederationService } from "../../dist/services/federation.js";
+import { GuardService } from "../../dist/services/guard.js";
 
 const url = process.env.SUPABASE_URL ?? "http://127.0.0.1:54321";
 const key = process.env.SUPABASE_KEY ?? "";
